+++ packages/cms/lib/modules/choices-guide-result-widgets/index.js
@@ -62,7 +62,7 @@ module.exports = {
 			widgets.forEach((widget) => {
         let apiUrl = self.apos.settings.getOption(req, 'apiUrl')
 			  widget.config = JSON.stringify(createConfig(widget, req.data, req.session.jwt, apiUrl, req.data.siteUrl + '/oauth/login?returnTo=' + encodeURIComponent(req.url), apiUrl + '/oauth/logout' ));
-        widget.openstadComponentsUrl = self.apos.settings.getOption(req, 'siteConfig').openstadComponentsUrl;
+        widget.openstadComponentsCdn = self.apos.settings.getOption(req, 'siteConfig').openstadComponentsCdn;
         const containerId = self.apos.utils.generateId();
         widget.containerId = containerId;
               widget.cssHelperClassesString = widget.cssHelperClasses ? widget.cssHelperClasses.join(' ') : '';