+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## UNRELEASED
+* Add create/update date to choices-guide results
+
 ## v0.14.0 (2020-03-17)
 * Add stats overview route for displaying statistics, currently used in react admin on CMS and Admin panel
 * Add isActive check on choices-guide