+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## UNRELEASED
 * Add create/update date to choices-guide results
+* Allow moderators to update role of users to member (but not to moderator or admin role)
 
 ## v0.14.0 (2020-03-17)
 * Add stats overview route for displaying statistics, currently used in react admin on CMS and Admin panel